feat: show reportable excerpt on hover
