+++ decidim-admin/app/views/decidim/admin/moderations/index.html.erb
@@ -40,7 +40,7 @@
               </td>
               <td>
                 <%=
-                  link_to t("models.moderation.fields.visit_url", scope: "decidim.moderations"), moderation.reportable.reported_content_url
+                  link_to t("models.moderation.fields.visit_url", scope: "decidim.moderations"), moderation.reportable.reported_content_url, data: { tooltip: true }, aria: { haspopup: true }, title: reported_content_excerpt_for(moderation.reportable)
                 %>
               </td>
               <td>
